new versions of all the files
particularly designed around the idea of working as a "gcode" reader
now. "gcode" will call pre-made functions and be read to tell the
plotter what to draw. A gcode creator and intepreter remains to be made
after the plotter itself works.

Also updated method for drawing a line between two points to better
allow bridging between points that are not in clear cut 45 / 90 degree
angles from eachother relative to the machine